separate wallet tests
