+++ tests/cmds/wallet/config.py
@@ -1,0 +1,3 @@
+from __future__ import annotations
+
+parallel = True
