+++ tests/cmds/wallet/test_wallet_check.py
